requirements.txt
openai-whisper
TTS
requests
